# Fix bulk modulus assignment and remove magic if from fixed stress porosity