Results for PoroElastic_Mandel_benchmark:

BEFORE
```
SinglePhaseFlow, number of time steps: 406
SinglePhaseFlow, number of successful nonlinear iterations: 5228
SinglePhaseFlow, number of successful linear iterations: 5228
SinglePhaseFlow, number of time step cuts: 0
SinglePhaseFlow, number of discarded nonlinear iterations: 0
SinglePhaseFlow, number of discarded linear iterations: 0
SinglePhaseFlow: apply solution time = 0.415142593 s (min), 0.415142593 s (max)
SinglePhaseFlow: assemble time = 12.163587556 s (min), 12.163587556 s (max)
SinglePhaseFlow: convergence check time = 0.296807929 s (min), 0.296807929 s (max)
SinglePhaseFlow: linear solver create time = 0.564239068 s (min), 0.564239068 s (max)
SinglePhaseFlow: linear solver setup time = 4.399641692 s (min), 4.399641692 s (max)
SinglePhaseFlow: linear solver solve time = 0.371110135 s (min), 0.371110135 s (max)
SinglePhaseFlow: linear solver total time = 5.47389087 s (min), 5.47389087 s (max)
SinglePhaseFlow: update state time = 0.296379135 s (min), 0.296379135 s (max)
lagsolve, number of time steps: 406
lagsolve, number of successful nonlinear iterations: 5227
lagsolve, number of successful linear iterations: 5227
lagsolve, number of time step cuts: 0
lagsolve, number of discarded nonlinear iterations: 0
lagsolve, number of discarded linear iterations: 0
lagsolve: apply solution time = 0.694335835 s (min), 0.694335835 s (max)
lagsolve: assemble time = 46.35469781 s (min), 46.35469781 s (max)
lagsolve: convergence check time = 0.55177106 s (min), 0.55177106 s (max)
lagsolve: linear solver create time = 8.371216151 s (min), 8.371216151 s (max)
lagsolve: linear solver setup time = 128.323329863 s (min), 128.323329863 s (max)
lagsolve: linear solver solve time = 9.865611708 s (min), 9.865611708 s (max)
lagsolve: linear solver total time = 148.527633954 s (min), 148.527633954 s (max)
lagsolve: update state time = 0.001346909 s (min), 0.001346909 s (max)
poroSolve, number of time steps: 406
poroSolve, number of successful nonlinear iterations: 5227
poroSolve, number of time step cuts: 0
poroSolve, number of discarded nonlinear iterations: 0
```

AFTER
```
SinglePhaseFlow, number of time steps: 406
SinglePhaseFlow, number of successful nonlinear iterations: 2648
SinglePhaseFlow, number of successful linear iterations: 2648
SinglePhaseFlow, number of time step cuts: 0
SinglePhaseFlow, number of discarded nonlinear iterations: 0
SinglePhaseFlow, number of discarded linear iterations: 0
SinglePhaseFlow: apply solution time = 0.190189186 s (min), 0.190189186 s (max)
SinglePhaseFlow: assemble time = 6.118371146 s (min), 6.118371146 s (max)
SinglePhaseFlow: convergence check time = 0.228266282 s (min), 0.228266282 s (max)
SinglePhaseFlow: linear solver create time = 0.279349996 s (min), 0.279349996 s (max)
SinglePhaseFlow: linear solver setup time = 2.202651688 s (min), 2.202651688 s (max)
SinglePhaseFlow: linear solver solve time = 0.181981649 s (min), 0.181981649 s (max)
SinglePhaseFlow: linear solver total time = 2.729876402 s (min), 2.729876402 s (max)
SinglePhaseFlow: update state time = 0.161967774 s (min), 0.161967774 s (max)
lagsolve, number of time steps: 406
lagsolve, number of successful nonlinear iterations: 2647
lagsolve, number of successful linear iterations: 2647
lagsolve, number of time step cuts: 0
lagsolve, number of discarded nonlinear iterations: 0
lagsolve, number of discarded linear iterations: 0
lagsolve: apply solution time = 0.332147426 s (min), 0.332147426 s (max)
lagsolve: assemble time = 22.885867387 s (min), 22.885867387 s (max)
lagsolve: convergence check time = 0.230638909 s (min), 0.230638909 s (max)
lagsolve: linear solver create time = 3.826605209 s (min), 3.826605209 s (max)
lagsolve: linear solver setup time = 64.340673516 s (min), 64.340673516 s (max)
lagsolve: linear solver solve time = 4.841161713 s (min), 4.841161713 s (max)
lagsolve: linear solver total time = 74.207833394 s (min), 74.207833394 s (max)
lagsolve: update state time = 0.0008579 s (min), 0.0008579 s (max)
poroSolve, number of time steps: 406
poroSolve, number of successful nonlinear iterations: 2647
poroSolve, number of time step cuts: 0
poroSolve, number of discarded nonlinear iterations: 0
```