@sohailwaziri 2023-10-18
Why is the outer convergence affected by the fix?