@paveltomin 2023-10-18
> Why is the outer convergence affected by the fix?

i am not 100% sure but behavior is different from the beginning because of that fixed-stress term.
before it was always ignored at first outer iteration.