@paveltomin 2023-10-18
> > > Why is the outer convergence affected by the fix?
> > 
> > 
> > i am not 100% sure but behavior is different from the beginning because of that fixed-stress term. before it was always ignored at first outer iteration.
> 
> It may also be a subtlety with the Mandel problem in particular. I have run some problems where it wasn't as sensitive.

Good point.
For PoroElastic_deadoil_3ph_baker_2d_sequential case it is 130 vs 132 outer iters.
For PoroElastic_hybridHexPrism_co2_sequential_3d it is 58 vs 58.
So you are probably right about Mandel.