@paveltomin 2023-10-20
> We should add comments in the code so that it's more clear which method are used by sequential runs and which by fully-coupled ones. It's not always easy to follow

You are right but I am afraid to fall into that rabbit hole in this PR...